refactor: Querydsl 기반의 Auction 조회 인터페이스를 domain 패키지로 이동
