+++ backend/ddang/src/main/java/com/ddang/ddang/auction/domain/dto/AuctionAndImageDto.java
@@ -1,4 +1,4 @@
-package com.ddang.ddang.auction.infrastructure.persistence.dto;
+package com.ddang.ddang.auction.domain.dto;
 
 import com.ddang.ddang.auction.domain.Auction;
 import com.ddang.ddang.image.domain.AuctionImage;
