+++ backend/ddang/src/main/java/com/ddang/ddang/auction/domain/repository/AuctionAndImageRepository.java
@@ -1,0 +1,9 @@
+package com.ddang.ddang.auction.domain.repository;
+
+import com.ddang.ddang.auction.domain.dto.AuctionAndImageDto;
+import java.util.Optional;
+
+public interface AuctionAndImageRepository {
+
+    Optional<AuctionAndImageDto> findDtoByAuctionId(final Long auctionId);
+}
