+++ backend/ddang/src/main/java/com/ddang/ddang/auction/infrastructure/persistence/JpaAuctionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.Query;
 
-public interface JpaAuctionRepository extends JpaRepository<Auction, Long>, QuerydslAuctionAndImageRepository {
+public interface JpaAuctionRepository extends JpaRepository<Auction, Long> {
 
     @Query("""
         SELECT a
