+++ backend/ddang/src/main/java/com/ddang/ddang/auction/infrastructure/persistence/dto/AuctionAndImageQueryProjectionDto.java
@@ -1,6 +1,7 @@
 package com.ddang.ddang.auction.infrastructure.persistence.dto;
 
 import com.ddang.ddang.auction.domain.Auction;
+import com.ddang.ddang.auction.domain.dto.AuctionAndImageDto;
 import com.ddang.ddang.image.domain.AuctionImage;
 import com.querydsl.core.annotations.QueryProjection;
